pyproject.toml
[project]
name = "theoria-agents"
version = "0.1.0"
description = "Multi-agent LLM system for generating high-quality theoretical physics dataset entries"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
theoria-generate = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
